[AllBundles] Added support for both symfony 3 and 4 for our overrides on the container access methods (#2097)
