+++ src/Kunstmaan/AdminBundle/Controller/BaseSettingsController.php
@@ -3,30 +3,16 @@
 namespace Kunstmaan\AdminBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpKernel\Kernel;
 
-class BaseSettingsController extends Controller
-{
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated
-     */
-    protected function get($id)
+if (Kernel::VERSION_ID < 40000) {
+    class_alias('Kunstmaan\AdminBundle\Controller\Symfony3BaseSettingsController', 'Kunstmaan\AdminBundle\Controller\BaseSettingsController');
+} else {
+    class_alias('Kunstmaan\AdminBundle\Controller\Symfony4BaseSettingsController', 'Kunstmaan\AdminBundle\Controller\BaseSettingsController');
+}
+
+if (false) {
+    class BaseSettingsController extends Controller
     {
-        @trigger_error('Getting services directly from the container is deprecated in KunstmaanAdminBundle 5.1 and will be removed in KunstmaanAdminBundle 6.0. Register your controllers as services and inject the necessary dependencies.', E_USER_DEPRECATED);
-
-        return parent::get($id);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated
-     */
-    protected function getParameter($name)
-    {
-        @trigger_error('Getting parameters directly from the container is deprecated in KunstmaanAdminBundle 5.1 and will be removed in KunstmaanAdminBundle 6.0. Register your controllers as services and inject the necessary parameters.', E_USER_DEPRECATED);
-
-        return parent::getParameter($name);
     }
 }
