+++ src/Kunstmaan/AdminBundle/Controller/Symfony4BaseSettingsController.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Kunstmaan\AdminBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+
+class Symfony4BaseSettingsController extends Controller
+{
+    /**
+     * {@inheritdoc}
+     *
+     * @deprecated
+     */
+    protected function get(string $id)
+    {
+        @trigger_error('Getting services directly from the container is deprecated in KunstmaanAdminBundle 5.1 and will be removed in KunstmaanAdminBundle 6.0. Register your controllers as services and inject the necessary dependencies.', E_USER_DEPRECATED);
+
+        return parent::get($id);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @deprecated
+     */
+    protected function getParameter(string $name)
+    {
+        @trigger_error('Getting parameters directly from the container is deprecated in KunstmaanAdminBundle 5.1 and will be removed in KunstmaanAdminBundle 6.0. Register your controllers as services and inject the necessary parameters.', E_USER_DEPRECATED);
+
+        return parent::getParameter($name);
+    }
+}
