+++ src/Kunstmaan/AdminListBundle/Controller/AdminListController.php
@@ -26,33 +26,8 @@ use Symfony\Component\PropertyAccess\PropertyAccess;
 /**
  * AdminListController
  */
-abstract class AdminListController extends Controller
+abstract class AdminListController extends ContainerDeprecationController
 {
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated
-     */
-    protected function get($id)
-    {
-        @trigger_error('Getting services directly from the container is deprecated in KunstmaanAdminListBundle 5.1 and will be removed in KunstmaanAdminListBundle 6.0. Register your controllers as services and inject the necessary dependencies.', E_USER_DEPRECATED);
-
-        return parent::get($id);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated
-     */
-    protected function getParameter($name)
-    {
-        @trigger_error('Getting parameters directly from the container is deprecated in KunstmaanAdminListBundle 5.1 and will be removed in KunstmaanAdminListBundle 6.0. Register your controllers as services and inject the necessary parameters.', E_USER_DEPRECATED);
-
-        return parent::getParameter($name);
-    }
-
-
     /**
      * You can override this method to return the correct entity manager when using multiple databases ...
      *
